+++ packages/checkout/core/src/providers/pix/pix.types.ts
@@ -1,10 +1,12 @@
+import { MalgaPaymentsItem } from '../../types/malga-payments-items.types'
+
 export interface PixAttributes {
   expiresIn: number
+  items?: MalgaPaymentsItem[] | null
 }
 
-export interface PaymentMethodPix {
+export interface PaymentMethodPix extends PixAttributes {
   paymentType: 'pix'
-  expiresIn: number
 }
 
 export interface PixConstructor {