+++ packages/checkout/core/src/services/charges/charges.utils.ts
@@ -44,7 +44,7 @@ export const formatFraudAnalysis = async (
   )
     return null
 
-  const parsedCustomer = formatCustomer(currentCustomer)
+  const parsedCustomer = formatCustomer(currentCustomer, true)
 
   const address = parsedCustomer.address && {
     ...parsedCustomer.address,
@@ -59,17 +59,23 @@ export const formatFraudAnalysis = async (
     delete address.streetNumber
   }
 
+  const phone = parsedCustomer?.phoneNumber?.trim()
+    ? { phone: parsedCustomer.phoneNumber }
+    : {}
+
   return {
     customer: {
       name: parsedCustomer.name,
       email: parsedCustomer.email,
-      phone: parsedCustomer.phoneNumber,
-      identityType: parsedCustomer.document.type.toUpperCase(),
-      identity: parsedCustomer.document.number,
+      ...(parsedCustomer.document && {
+        identityType: parsedCustomer.document.type,
+        identity: parsedCustomer.document.number,
+      }),
       ...(currentCustomer.address && {
         deliveryAddress: address,
         billingAddress: address,
       }),
+      ...phone,
       ...browser,
     },
     cart: {