+++ packages/checkout/core/src/services/customers/customers.utils.ts
@@ -32,8 +32,13 @@ export const formatCustomerAddress = (
 
 export const formatCustomerDocument = (
   customerDocument?: Customer['document'],
+  isFraudAnalysis: boolean = false,
 ) => {
   if (!customerDocument || customerDocument.type === 'noDocument') {
+    if (isFraudAnalysis) {
+      return null
+    }
+
     return {
       document: {
         type: 'noDocument',
@@ -50,22 +55,26 @@ export const formatCustomerDocument = (
 
   return {
     document: {
-      type: customerDocument.type,
+      type: isFraudAnalysis
+        ? customerDocument.type.toUpperCase()
+        : customerDocument.type.toLowerCase(),
       number: documentNumber,
       country: customerDocument.country,
     },
   }
 }
 
-export const formatPayload = (customer: Customer) => {
+export const formatPayload = (customer: Customer, isFraudAnalysis: boolean = false) => {
   const address = formatCustomerAddress(customer.address)
-  const document = formatCustomerDocument(customer.document)
+  const document = formatCustomerDocument(customer.document, isFraudAnalysis)
 
   return {
     ...address,
     ...document,
+    phoneNumber: customer?.phoneNumber
+      ? cleanTextOnlyNumbers(customer.phoneNumber)
+      : ' ',
     name: customer.name,
     email: customer.email,
-    phoneNumber: cleanTextOnlyNumbers(customer.phoneNumber),
   }
 }