+++ packages/common/src/components/checkout-button/checkout-button.tsx
@@ -68,7 +68,7 @@ export class CheckoutButton implements ComponentInterface {
   render() {
     return (
       <Host
-        class={{ 'checkout-button__container--full-width': this.fullWidth }}
+        style={{ pointerEvents: this.disabled ? 'none' : undefined }}
         aria-disabled={this.disabled ? 'true' : null}
       >
         <button